Comparative TFBS scan
comparative_scan_rtfbs function uses a PWM to scan for TFBS in two sets (positive and negative) with either a FDR or log odds ratio threshold. Sequences are split into GC content quantiles and a background model is build for each. If an FDR threshold is requested, a simulated background sequence is generated to estimate the log odds ratio threshold that achieves the desired FDR (per GC content quantile).

comparative_scanDb_rtfbs does the same, but scans over a set of PWMs, returning a summary table at the end.

Additional functions:
- write.starchbed :: pipe a BED data.frame through starch
- comparative_scan_rtfbs.ms :: core function used by both the above functions